+++ src/components/modals/RegModal.tsx
